Update HTML processing
