+++ src/helpers/index.ts
@@ -1,1 +1,3 @@
 export * from './handlebars';
+export * from './html';
+export * from './variable';
